[NEW RBO] Split logical property propagation
