+++ ydb/core/kqp/opt/rbo/kqp_operator.cpp
@@ -35,22 +35,6 @@ TString FormatSortElements(const TVector<TSortElement>& sortElements) {
     return result;
 }
 
-bool AddInfoUnit(TInfoUnitSet& target, const TInfoUnit& iu) {
-    return target.insert(iu).second;
-}
-
-void AddInfoUnits(TInfoUnitSet& target, const TVector<TInfoUnit>& ius) {
-    for (const auto& iu : ius) {
-        AddInfoUnit(target, iu);
-    }
-}
-
-TInfoUnitSet MakeInfoUnitSet(const TVector<TInfoUnit>& ius) {
-    TInfoUnitSet result;
-    AddInfoUnits(result, ius);
-    return result;
-}
-
 TInfoUnit RenameInfoUnit(const TInfoUnit& iu, const THashMap<TInfoUnit, TInfoUnit, TInfoUnit::THashFunction>& renameMap) {
     const auto it = renameMap.find(iu);
     return it == renameMap.end() ? iu : it->second;
@@ -163,21 +147,6 @@ NJson::TJsonValue IOperator::ToJson(ui32 explainFlags)
     return res;
 }
 
-void IOperator::PropagateLiveness(ILivenessContext& ctx) {
-    Y_UNUSED(ctx);
-}
-
-void IUnaryOperator::PropagateLiveness(ILivenessContext& ctx) {
-    const TInfoUnitSet liveOut = ctx.GetLiveOut(this);
-    TInfoUnitSet inputLive;
-    for (const auto& iu : GetInput()->GetOutputIUs()) {
-        if (liveOut.contains(iu)) {
-            AddInfoUnit(inputLive, iu);
-        }
-    }
-    ctx.AddLiveColumns(GetInput(), inputLive);
-}
-
 /**
  * EmptySource operator methods
  */
@@ -227,10 +196,6 @@ TVector<TInfoUnit> TOpRead::GetOutputIUs() {
         return *outputIUs;
     }
     return OutputIUs;
-}
-
-void TOpRead::PropagateLiveness(ILivenessContext& ctx) {
-    Y_UNUSED(ctx);
 }
 
 bool TOpRead::NeedsMap() const {
@@ -489,30 +454,6 @@ TVector<TInfoUnit> TOpMap::GetSubplanIUs(TPlanProps& props) {
     return res;
 }
 
-void TOpMap::PropagateLiveness(ILivenessContext& ctx) {
-    const TInfoUnitSet liveOut = ctx.GetLiveOut(this);
-    auto input = GetInput();
-    TInfoUnitSet inputLive;
-    TInfoUnitSet renameSources;
-
-    for (const auto& mapElement : MapElements) {
-        if (mapElement.IsRename()) {
-            renameSources.insert(mapElement.GetRename());
-        }
-        // Keep dependencies of every current map expression live so local pruning
-        // cannot remove producer columns before the dead consumer expression is gone.
-        ctx.AddExpressionDeps(mapElement.GetExpression(), inputLive);
-    }
-
-    for (const auto& iu : input->GetOutputIUs()) {
-        if (!renameSources.contains(iu) && liveOut.contains(iu)) {
-            AddInfoUnit(inputLive, iu);
-        }
-    }
-
-    ctx.AddLiveColumns(input, inputLive);
-}
-
 // Returns explicit renames as pairs of <to, from>
 
 TVector<std::pair<TInfoUnit, TInfoUnit>> TOpMap::GetRenames() const {
@@ -763,12 +704,6 @@ TVector<TInfoUnit> TOpFilter::GetUsedIUs(TPlanProps& props) {
     return FilterExpr.GetInputIUs(false, true);
 }
 
-void TOpFilter::PropagateLiveness(ILivenessContext& ctx) {
-    TInfoUnitSet inputLive = ctx.GetLiveOut(this);
-    ctx.AddExpressionDeps(FilterExpr, inputLive);
-    ctx.AddLiveColumns(GetInput(), inputLive);
-}
-
 TVector<TInfoUnit> TOpFilter::GetSubplanIUs(TPlanProps& props) {
     TVector<TInfoUnit> res;
     for (const auto& iu : GetFilterIUs(props)) {
@@ -867,57 +802,6 @@ TVector<std::reference_wrapper<TExpression>> TOpJoin::GetExpressions() {
         result.push_back(expr);
     }
     return result;
-}
-
-void TOpJoin::PropagateLiveness(ILivenessContext& ctx) {
-    const TInfoUnitSet liveOut = ctx.GetLiveOut(this);
-    const auto leftInput = GetLeftInput();
-    const auto rightInput = GetRightInput();
-    const auto leftOutput = MakeInfoUnitSet(leftInput->GetOutputIUs());
-    const auto rightOutput = MakeInfoUnitSet(rightInput->GetOutputIUs());
-
-    TInfoUnitSet leftLive;
-    TInfoUnitSet rightLive;
-
-    const bool outputsLeft = JoinKind != "RightOnly" && JoinKind != "RightSemi";
-    const bool outputsRight = JoinKind != "LeftOnly" && JoinKind != "LeftSemi";
-
-    if (outputsLeft) {
-        for (const auto& iu : leftOutput) {
-            if (liveOut.contains(iu)) {
-                AddInfoUnit(leftLive, iu);
-            }
-        }
-    }
-
-    if (outputsRight) {
-        for (const auto& iu : rightOutput) {
-            if (liveOut.contains(iu)) {
-                AddInfoUnit(rightLive, iu);
-            }
-        }
-    }
-
-    for (const auto& [leftKey, rightKey] : JoinKeys) {
-        AddInfoUnit(leftLive, leftKey);
-        AddInfoUnit(rightLive, rightKey);
-    }
-
-    for (const auto& filter : JoinFilters) {
-        TInfoUnitSet filterDeps;
-        ctx.AddExpressionDeps(filter, filterDeps);
-        for (const auto& iu : filterDeps) {
-            if (leftOutput.contains(iu)) {
-                AddInfoUnit(leftLive, iu);
-            }
-            if (rightOutput.contains(iu)) {
-                AddInfoUnit(rightLive, iu);
-            }
-        }
-    }
-
-    ctx.AddLiveColumns(leftInput, leftLive);
-    ctx.AddLiveColumns(rightInput, rightLive);
 }
 
 TString GetJoinAlgoName(NKqp::EJoinAlgoType joinAlgo) {
@@ -1047,12 +931,6 @@ TVector<TInfoUnit> TOpUnionAll::GetOutputIUs() {
     return commonOutput.empty() ? leftOutput : commonOutput;
 }
 
-void TOpUnionAll::PropagateLiveness(ILivenessContext& ctx) {
-    const TInfoUnitSet liveOut = ctx.GetLiveOut(this);
-    ctx.AddLiveColumns(GetLeftInput(), liveOut);
-    ctx.AddLiveColumns(GetRightInput(), liveOut);
-}
-
 TString TOpUnionAll::ToString(TExprContext& ctx) {
     Y_UNUSED(ctx); 
     return "UnionAll";
@@ -1100,15 +978,6 @@ TVector<TInfoUnit> TOpLimit::GetOutputIUs() {
         return *outputIUs;
     }
     return GetInput()->GetOutputIUs();
-}
-
-void TOpLimit::PropagateLiveness(ILivenessContext& ctx) {
-    TInfoUnitSet inputLive = ctx.GetLiveOut(this);
-    ctx.AddExpressionDeps(LimitCond, inputLive);
-    if (OffsetCond) {
-        ctx.AddExpressionDeps(*OffsetCond, inputLive);
-    }
-    ctx.AddLiveColumns(GetInput(), inputLive);
 }
 
 void TOpLimit::RenameIUs(const THashMap<TInfoUnit, TInfoUnit, TInfoUnit::THashFunction>& renameMap, TExprContext& ctx,
@@ -1180,17 +1049,6 @@ TVector<TInfoUnit> TOpSort::GetUsedIUs(TPlanProps& props) {
         result.push_back(element.SortColumn);
     }
     return result;
-}
-
-void TOpSort::PropagateLiveness(ILivenessContext& ctx) {
-    TInfoUnitSet inputLive = ctx.GetLiveOut(this);
-    for (const auto& sortElement : SortElements) {
-        AddInfoUnit(inputLive, sortElement.SortColumn);
-    }
-    if (LimitCond) {
-        ctx.AddExpressionDeps(*LimitCond, inputLive);
-    }
-    ctx.AddLiveColumns(GetInput(), inputLive);
 }
 
 void TOpSort::RenameIUs(const THashMap<TInfoUnit, TInfoUnit, TInfoUnit::THashFunction> &renameMap, TExprContext &ctx, const THashSet<TInfoUnit, TInfoUnit::THashFunction> &stopList) {
@@ -1304,15 +1162,6 @@ TVector<TInfoUnit> TOpAggregate::GetUsedIUs(TPlanProps& props) {
         usedIUs.push_back(aggTraits.OriginalColName);
     }
     return usedIUs;
-}
-
-void TOpAggregate::PropagateLiveness(ILivenessContext& ctx) {
-    TInfoUnitSet inputLive;
-    AddInfoUnits(inputLive, KeyColumns);
-    for (const auto& traits : AggregationTraitsList) {
-        AddInfoUnit(inputLive, traits.OriginalColName);
-    }
-    ctx.AddLiveColumns(GetInput(), inputLive);
 }
 
 void TOpAggregate::RenameIUs(const THashMap<TInfoUnit, TInfoUnit, TInfoUnit::THashFunction>& renameMap, TExprContext& ctx,
@@ -1437,12 +1286,6 @@ void TOpCBOTree::RebuildChildren() {
 
             Children.push_back(child);
         }
-    }
-}
-
-void TOpCBOTree::PropagateLiveness(ILivenessContext& ctx) {
-    for (const auto& child : Children) {
-        ctx.AddLiveColumns(child, child->GetOutputIUs());
     }
 }
